+++ CHANGELOG.md
@@ -7,8 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-### Added
+### Changed
+- Renamed `postgres.py` to `pgjinja.py` for better module name consistency
 
+### Added
 ## [0.1.0] - 2025-03-26
 
 ### Added